refactor web3 rpc
